+++ gradio/__init__.py
@@ -10,6 +10,8 @@ from gradio.blocks import Blocks
 from gradio.components import (
     HTML,
     JSON,
+    AnnotatedImage,
+    Annotatedimage,
     Audio,
     BarPlot,
     Button,